pyproject.toml
[project]
name = "gsoc"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "google-generativeai>=0.8.5",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "typer>=0.16.0",
]
